// src/global.d.ts
  interface breadcrumb {
    /**
     * 是否开启面包屑导航
     * @默认值 `true`
     */
    enable?: boolean
  }
// ...
  interface RouteMeta {
    title: string | Function
    i18n?: string
    icon?: string
    activeIcon?: string
    defaultOpened?: boolean
    auth?: string | string[]
    sidebar?: boolean
    breadcrumb?: boolean
    activeMenu?: string
    cache?: boolean | string | string[]
    breadcrumbNeste?: Route.breadcrumb[]
  }
// ...
declare namespace Route {
  interface recordRaw extends Omit<RouteRecordRaw, 'meta' | 'children'> {
    // meta: {
    //   title?: string | Function
    //   icon?: string
    //   auth?: string | string[]
    //   hide?: boolean
    //   sidebar?: boolean
    // }
    icon?: string
    breadcrumbNeste?: breadcrumb[];
    breadcrumb?: any;
    title?: any;
    auth?: string | string[]
    meta?: RouteMeta
    children?: recordRaw[]
    sidebar?: boolean
    // beforeRouterEnter?:Function
  }
  interface recordMainRaw {
    meta: {
      title?: string | Function
      icon?: string
      auth?: string | string[]
      hide?: boolean
      sidebar?: boolean
    }
    icon?: string
    auth?: string | string[]
    hide?: boolean
    title?: string | Function
    children: recordRaw[]
    redirect?:string
    path?:String
  }
  interface breadcrumb {
    path: string
    title: string | Function
    hide: boolean
  }
}